Добавлена возможность редактирования проекта + добавление нескольких ТМов
TODO:
1) Отслеживание промежутков размещения отчетов
2) Добавление ошибок
3) Страница скачивания PDF с отчетом
4) Скачивание прикрепленных файлов

+++ src/main/java/report/pflb/ProjectReport/service/userService.java
@@ -24,6 +24,6 @@ public class userService{
 
     public List<User> getAllFromDatabase(){return userRepository.findAll();}
 
-    public User findById(int id){return userRepository.findById(id);}
+    public User findById(long id){return userRepository.findById(id);}
     public User findByLogin(String login){return userRepository.findByLogin(login);}
 }
